Computer_Vision/Similarity_with_the_actor/requirements.txt
PyYAML==6.0
bing_image_downloader==1.1.1
numpy==1.22.1
pandas==1.3.5
scikit-learn==1.0.2
tqdm==4.62.3
Pillow==9.0.0
